+++ src/components/Match/Dday.tsx
@@ -10,9 +10,9 @@ const Dday = (props: JoinDataProps) => {
   const abs = Math.abs(c);
 
   if (c < 1 && a.format('YYYY-MM-DD') !== now.format('YYYY-MM-DD')) {
-    return <p className='w-[5rem] h-7 text-[#38393C]'>(D-{abs + 1})</p>;
+    return <p className='w-[5rem] h-7 text-[#38393C] font-SD'>(D-{abs + 1})</p>;
   } else if (c < 1 && a.format('YYYY-MM-DD') === now.format('YYYY-MM-DD')) {
-    return <p className='w-[5rem] h-7 text-[#38393C]'>(D-DAY)</p>;
+    return <p className='w-[5rem] h-7 text-[#38393C] font-SD'>(D-DAY)</p>;
   } else {
     return <></>;
   }
