+++ src/hooks/queries/useGuestPostList.ts
@@ -1,11 +1,32 @@
 import { useQuery } from '@tanstack/react-query';
 import { queryKeys } from '../../shared/constant/queryKeys';
-import { PostDataProps } from '../../typings';
+import { PostDataProps, JoinDataProps } from '../../typings';
 import { apis } from '../../apis';
 
-export function useGetPostList<T>(postId: number) {
+export function useGetPostList(postId: number) {
   const { data: res, isLoading } = useQuery([queryKeys.GUESTLIST, postId], () => apis.getforGuestPostList(postId));
   const guestData: PostDataProps = res?.data;
 
-  return { guestData, isLoading };
+  const drill: JoinDataProps = {
+    data: {
+      owner: guestData?.owner,
+      matchStatus: guestData?.matchStatus,
+      postId: postId,
+      player: guestData?.player,
+      profileUrl: guestData?.profileImgUrl,
+      nickName: guestData?.nickname,
+      matchDeadline: guestData?.matchDeadline,
+      lat: guestData?.lat,
+      lng: guestData?.lng,
+      address: guestData?.address,
+      imgpaths: guestData?.imgpaths,
+      imgurls: guestData?.imgurls,
+      subjectValue: guestData?.subjectValue,
+      subject: guestData?.subject,
+      title: guestData?.title,
+      content: guestData?.content,
+    },
+  };
+
+  return { guestData, isLoading, drill };
 }
